wrap App in PaperProvider
